# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Raising Talents API",
    "version": "0.0.0",
    "description": "Generated from @rt/contracts. Errors use RFC 9457 problem details; choose behaviour by the `code` field, never the message."
  },
  "servers": [
    {
      "url": "https://api.raisingtalents.app",
      "description": "Production"
    },
    {
      "url": "https://api.staging.raisingtalents.app",
      "description": "Staging"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
...
  "paths": {
...
    "/v1/me/talent-profile": {
      "get": {
        "operationId": "talentProfile_getMine",
        "summary": "The signed-in talent's profile, with what is still missing",
        "tags": [
          "Talent profiles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Version to send back in If-Match.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyTalentProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. Codes: VALIDATION_FAILED.",
            "x-error-codes": [
              "VALIDATION_FAILED"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in. Codes: UNAUTHENTICATED, SESSION_EXPIRED.",
            "x-error-codes": [
              "UNAUTHENTICATED",
              "SESSION_EXPIRED"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed. Codes: WRONG_ROLE.",
            "x-error-codes": [
              "WRONG_ROLE"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Codes: NOT_FOUND.",
            "x-error-codes": [
              "NOT_FOUND"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong. Codes: INTERNAL.",
            "x-error-codes": [
              "INTERNAL"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "talentProfile_updateMine",
        "summary": "Create or update the talent profile, one onboarding step at a time",
        "tags": [
          "Talent profiles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "The ETag from the last read. Required once the resource exists; a stale value returns 412.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTalentProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Version to send back in If-Match.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyTalentProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. Codes: VALIDATION_FAILED.",
            "x-error-codes": [
              "VALIDATION_FAILED"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in. Codes: UNAUTHENTICATED, SESSION_EXPIRED.",
            "x-error-codes": [
              "UNAUTHENTICATED",
              "SESSION_EXPIRED"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed. Codes: WRONG_ROLE, EMAIL_NOT_VERIFIED.",
            "x-error-codes": [
              "WRONG_ROLE",
              "EMAIL_NOT_VERIFIED"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict. Codes: HANDLE_TAKEN.",
            "x-error-codes": [
              "HANDLE_TAKEN"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "412": {
            "description": "Stale version. Codes: PRECONDITION_FAILED.",
            "x-error-codes": [
              "PRECONDITION_FAILED"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request cannot be completed. Codes: HANDLE_INVALID, UNKNOWN_TAXONOMY.",
            "x-error-codes": [
              "HANDLE_INVALID",
              "UNKNOWN_TAXONOMY"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "428": {
            "description": "If-Match header required. Codes: PRECONDITION_REQUIRED.",
            "x-error-codes": [
              "PRECONDITION_REQUIRED"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong. Codes: INTERNAL.",
            "x-error-codes": [
              "INTERNAL"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/talents/{handle}": {
      "get": {
        "operationId": "talents_getByHandle",
        "summary": "A complete talent profile, as agents see it",
        "tags": [
          "Talent profiles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicTalentProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. Codes: VALIDATION_FAILED.",
            "x-error-codes": [
              "VALIDATION_FAILED"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in. Codes: UNAUTHENTICATED, SESSION_EXPIRED.",
            "x-error-codes": [
              "UNAUTHENTICATED",
              "SESSION_EXPIRED"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Codes: NOT_FOUND.",
            "x-error-codes": [
              "NOT_FOUND"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong. Codes: INTERNAL.",
            "x-error-codes": [
              "INTERNAL"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FieldProblem": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ],
        "additionalProperties": false
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMITED",
              "INTERNAL",
              "EMAIL_ALREADY_REGISTERED",
              "UNDER_MINIMUM_AGE",
              "WEAK_PASSWORD",
              "INVALID_CREDENTIALS",
              "ACCOUNT_SUSPENDED",
              "ACCOUNT_BANNED",
              "SESSION_EXPIRED",
              "SESSION_REVOKED",
              "VERIFICATION_CODE_INVALID",
              "VERIFICATION_CODE_EXPIRED",
              "VERIFICATION_ATTEMPTS_EXCEEDED",
              "VERIFICATION_RESEND_TOO_SOON",
              "EMAIL_ALREADY_VERIFIED",
              "EMAIL_NOT_VERIFIED",
              "ROLE_ALREADY_LOCKED",
              "PRECONDITION_REQUIRED",
              "PRECONDITION_FAILED",
              "WRONG_ROLE",
              "HANDLE_TAKEN",
              "HANDLE_INVALID",
              "UNKNOWN_TAXONOMY"
            ]
          },
          "detail": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldProblem"
            }
          },
          "retryAfterSeconds": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "additionalProperties": false
      },
...
      "NamedRef": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name"
        ],
        "additionalProperties": false
      },
      "UpdateTalentProfileRequest": {
        "type": "object",
        "properties": {
          "handle": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "bio": {
            "type": "string",
            "maxLength": 1000
          },
          "categorySlug": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "subcategorySlugs": {
            "maxItems": 5,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 64,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          },
          "skillSlugs": {
            "maxItems": 15,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 64,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          },
          "citySlug": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "female",
                  "male",
                  "non_binary"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "genderSearchable": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "MyTalentProfile": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "handle": {
            "type": "string"
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "bio": {
            "type": "string"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NamedRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "subcategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedRef"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedRef"
            }
          },
          "city": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "slug": {
                    "type": "string",
                    "maxLength": 64,
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                  },
                  "name": {
                    "type": "string"
                  },
                  "countryCode": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 2
                  }
                },
                "required": [
                  "slug",
                  "name",
                  "countryCode"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "female",
                  "male",
                  "non_binary"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "genderSearchable": {
            "type": "boolean"
          },
          "avatarMediaId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ]
          },
          "isComplete": {
            "type": "boolean"
          },
          "missing": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "displayName",
                "category",
                "subcategories",
                "city",
                "bio",
                "avatar"
              ]
            }
          },
          "version": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "userId",
          "handle",
          "displayName",
          "bio",
          "category",
          "subcategories",
          "skills",
          "city",
          "gender",
          "genderSearchable",
          "avatarMediaId",
          "isComplete",
          "missing",
          "version",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "PublicTalentProfile": {
        "type": "object",
        "properties": {
          "handle": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/NamedRef"
          },
          "subcategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedRef"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedRef"
            }
          },
          "city": {
            "type": "object",
            "properties": {
              "slug": {
                "type": "string",
                "maxLength": 64,
                "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
              },
              "name": {
                "type": "string"
              },
              "countryCode": {
                "type": "string",
                "minLength": 2,
                "maxLength": 2
              }
            },
            "required": [
              "slug",
              "name",
              "countryCode"
            ],
            "additionalProperties": false
          },
          "ageYears": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "female",
                  "male",
                  "non_binary"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "verified": {
            "type": "boolean"
          },
          "avatarMediaId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "handle",
          "displayName",
          "bio",
          "category",
          "subcategories",
          "skills",
          "city",
          "ageYears",
          "gender",
          "verified",
          "avatarMediaId"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "15-minute access token from sign-in, sign-up or refresh."
      }
    }
  }
}